+++ packages/forms/src/FormSkeleton.test.js
@@ -1,21 +1,19 @@
 import React from 'react';
-import { shallow } from 'enzyme';
-import Skeleton from '@talend/react-components/lib/Skeleton';
+
+import { render, screen } from '@testing-library/react';
+
 import FormSkeleton from './FormSkeleton';
 
 describe('FormSkeleton', () => {
 	it('should render skeleton', () => {
-		const wrapper = shallow(<FormSkeleton />);
-		expect(wrapper.find(Skeleton).length).toBe(10);
-	});
-
-	it('should render skeleton  without its buttons', () => {
-		const wrapper = shallow(<FormSkeleton displayMode="text" />);
-		expect(wrapper.find(Skeleton).length).toBe(8);
+		render(<FormSkeleton />);
+		expect(screen.getByTestId('form.skeleton')).toHaveAttribute('aria-busy', 'true');
+		expect(screen.getByTestId('form.skeleton.fields')).toBeInTheDocument();
+		expect(screen.getByTestId('form.skeleton.buttons')).toBeInTheDocument();
 	});
 
 	it('should render skeleton without actions', () => {
-		const wrapper = shallow(<FormSkeleton actions={[]} />);
-		expect(wrapper.find(Skeleton).length).toBe(8);
+		render(<FormSkeleton actions={[]} />);
+		expect(screen.queryByTestId('form.skeleton.buttons')).not.toBeInTheDocument();
 	});
 });