+++ packages/containers/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @talend/react-containers
 
+## 7.2.3
+
+### Patch Changes
+
+- 1770e701b: fix: include peerDependencies in UMD manifest
+- Updated dependencies [1770e701b]
+  - @talend/react-cmf@7.1.1
+  - @talend/react-components@7.0.1
+  - @talend/react-forms@7.3.2
+
+## 7.2.2
+
+### Patch Changes
+
+- Updated dependencies [9a581a4bc]
+- Updated dependencies [c8fbf0130]
+- Updated dependencies [1160ec1f6]
+  - @talend/react-components@7.0.0
+  - @talend/react-forms@7.3.1
+
 ## 7.2.1
 
 ### Patch Changes